+++ pytorch_lightning/plugins/precision/precision_plugin.py
@@ -110,6 +110,28 @@ class PrecisionPlugin(CheckpointHooks):
         """
         tensor.backward(*args, **kwargs)
 
+    def _wrap_closure(
+        self,
+        model: Union["pl.LightningModule", Module],
+        optimizer: Optimizer,
+        optimizer_idx: int,
+        lambda_closure: Callable[[], Any],
+    ) -> Callable[[], Any]:
+        """This double-closure allows makes sure the ``lambda_closure`` is executed before the
+        ``on_before_optimizer_step`` hook is called.
+
+        The closure (generally) runs ``backward`` so this allows inspecting gradients in this hook. This structure is
+        consistent with the ``PrecisionPlugin`` subclasses that cannot pass ``optimizer.step(lambda_closure)`` directly.
+        """
+
+        def inner() -> Any:
+            closure_result = lambda_closure()
+            if isinstance(model, pl.LightningModule):
+                model.trainer.call_hook("on_before_optimizer_step", optimizer, optimizer_idx)
+            return closure_result
+
+        return inner
+
     def optimizer_step(
         self,
         model: Union["pl.LightningModule", Module],
@@ -119,12 +141,11 @@ class PrecisionPlugin(CheckpointHooks):
         **kwargs: Any,
     ) -> None:
         """Hook to run the optimizer step."""
-        if isinstance(model, pl.LightningModule):
-            model.trainer.call_hook("on_before_optimizer_step", optimizer, optimizer_idx)
-        optimizer.step(closure=lambda_closure, **kwargs)
+        closure = self._wrap_closure(model, optimizer, optimizer_idx, lambda_closure)
+        optimizer.step(closure=closure, **kwargs)
 
     def _track_grad_norm(self, trainer: "pl.Trainer") -> None:
-        if float(trainer.track_grad_norm) == -1:
+        if trainer.track_grad_norm == -1:
             return
         grad_norm_dict = grad_norm(trainer.lightning_module, trainer.track_grad_norm, trainer.logger.group_separator)
         if grad_norm_dict: