+++ pytorch_lightning/plugins/precision/tpu.py
@@ -34,9 +34,8 @@ class TPUPrecisionPlugin(PrecisionPlugin):
         lambda_closure: Callable[[], Any],
         **kwargs: Any
     ) -> None:
-        if isinstance(model, pl.LightningModule):
-            model.trainer.call_hook("on_before_optimizer_step", optimizer, optimizer_idx)
-        closure_result = xm.optimizer_step(optimizer, optimizer_args={"closure": lambda_closure, **kwargs})
+        closure = self._wrap_closure(model, optimizer, optimizer_idx, lambda_closure)
+        closure_result = xm.optimizer_step(optimizer, optimizer_args={"closure": closure, **kwargs})
         skipped_backward = closure_result is None
         # in manual optimization, the closure does not return a value
         if isinstance(model, pl.LightningModule) and model.automatic_optimization and skipped_backward: